mad props to junior for pointing out allocate_user_counter is per-user, not global.
git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@8518 1dfba125-df10-0410-8a77-af1a8d6d074f

+++ cgi-bin/LJ/Subscription.pm
@@ -181,7 +181,7 @@ sub delete {
 
     my @subs = $self->corresponding_subs;
     foreach my $subscr (@subs) {
-        $u->do("DELETE FROM subs WHERE subid=?", undef, $subscr->id);
+        $u->do("DELETE FROM subs WHERE subid=? AND userid=?", undef, $subscr->id, $u->id);
     }
 
     # delete from cache in user
